tests/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --emit-index-url --no-strip-extras tests/requirements.in -o tests/requirements.txt
--index-url https://blackboard.jfrog.io/artifactory/api/pypi/fnds-pypi/simple

bb-ent-data-services-shared-core==2.6.0
    # via bb-ent-data-services-shared-tests
bb-ent-data-services-shared-tests==2.6.0
    # via -r tests/requirements.in
bb-fnds-pyfnds==0.78.0
    # via bb-fnds-pyfnds-pytest
bb-fnds-pyfnds-pytest==0.78.0
    # via
    #   -r tests/requirements.in
    #   bb-ent-data-services-shared-tests
boto3==1.37.22
    # via
    #   moto
    #   requests-aws-sign
botocore==1.37.22
    # via
    #   boto3
    #   moto
    #   s3transfer
certifi==2025.1.31
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via requests
cryptography==44.0.2
    # via moto
dnspython==2.7.0
    # via bb-fnds-pyfnds
idna==3.10
    # via requests
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via moto
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonpath-ng==1.7.0
    # via bb-fnds-pyfnds
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
moto==5.1.1
    # via -r tests/requirements.in
packaging==24.2
    # via pytest
pluggy==1.5.0
    # via pytest
ply==3.11
    # via jsonpath-ng
pycparser==2.22
    # via cffi
pydantic==1.10.21
    # via bb-fnds-pyfnds
pyjwt==2.10.1
    # via -r tests/requirements.in
pytest==8.3.5
    # via
    #   bb-ent-data-services-shared-tests
    #   bb-fnds-pyfnds-pytest
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   moto
pytz==2025.2
    # via -r tests/requirements.in
pyyaml==6.0.2
    # via responses
requests==2.32.3
    # via
    #   bb-ent-data-services-shared-tests
    #   moto
    #   requests-aws-sign
    #   requests-mock
    #   responses
requests-aws-sign==0.1.6
    # via
    #   bb-ent-data-services-shared-tests
    #   bb-fnds-pyfnds
requests-mock==1.12.1
    # via -r tests/requirements.in
responses==0.25.7
    # via moto
s3transfer==0.11.4
    # via boto3
six==1.17.0
    # via python-dateutil
typing-extensions==4.13.0
    # via
    #   bb-fnds-pyfnds
    #   pydantic
urllib3==2.3.0
    # via
    #   botocore
    #   requests
    #   responses
werkzeug==3.1.3
    # via moto
xmltodict==0.14.2
    # via moto
